+++ snake-ai-sfml/game_core/include/Utilities.h
@@ -13,4 +13,4 @@ void DrawLine(sf::RenderWindow* window, float x1, float y1, float x2, float y2, 
 
 void DrawRectangle(sf::RenderWindow* window, float left, float top, float width, float height, sf::Color color);
 
-void print_ai_inputs(vector<int> inputs);
+void print_ai_inputs(vector<float> inputs);
